+++ supervisely/nn/benchmark/visualization/vis_metrics/overall_error_analysis.py
@@ -1,0 +1,109 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, List
+
+from supervisely.nn.benchmark.cv_tasks import CVTask
+from supervisely.nn.benchmark.visualization.vis_metric_base import MetricVis
+from supervisely.nn.benchmark.visualization.vis_widgets import Widget
+
+if TYPE_CHECKING:
+    from supervisely.nn.benchmark.visualization.visualizer import Visualizer
+
+
+class OverallErrorAnalysis(MetricVis):
+
+    def __init__(self, loader: Visualizer) -> None:
+        super().__init__(loader)
+        self.cv_tasks: List[CVTask] = [CVTask.SEGMENTATION.value]
+
+    def get_figure(self, widget: Widget.Chart):  # -> Optional[go.Figure]:
+        import plotly.graph_objects as go
+        from plotly.subplots import make_subplots
+
+        fig = make_subplots(
+            rows=1,
+            cols=3,
+            subplot_titles=(
+                "Basic segmentation metrics",
+                "Intersection & Error over Union",
+                "Renormalized Error over Union",
+            ),
+            specs=[[{"type": "polar"}, {"type": "domain"}, {"type": "xy"}]],
+        )
+
+        # first subplot
+        categories = [
+            "mPixel accuracy",
+            "mPrecision",
+            "mRecall",
+            "mF1-score",
+            "mIoU",
+            "mBoundaryIoU",
+            "mPixel accuracy",
+        ]
+        values = [64, 60.4, 52, 51.4, 37.9, 20.5, 64]
+        trace_1 = go.Scatterpolar(
+            mode="lines+text",
+            r=values,
+            theta=categories,
+            fill="toself",
+            fillcolor="cornflowerblue",
+            line_color="blue",
+            opacity=0.6,
+            text=[64, 60.4, 52, 51.4, 37.9, 20.5, 64],
+            textposition=[
+                "bottom right",
+                "top center",
+                "top center",
+                "middle left",
+                "bottom center",
+                "bottom right",
+                "bottom right",
+            ],
+            textfont=dict(color="blue"),
+        )
+        fig.add_trace(trace_1, row=1, col=1)
+
+        # second subplot
+        labels = ["mIoU", "mBoundaryEoU", "mExtentEoU", "mSegmentEoU"]
+        values = [37.9, 13.1, 25.8, 23.2]
+        trace_2 = go.Pie(
+            labels=labels,
+            values=values,
+            hole=0.5,
+            textposition="outside",
+            textinfo="percent+label",
+            marker=dict(colors=["cornflowerblue", "moccasin", "lightgreen", "orangered"]),
+        )
+        fig.add_trace(trace_2, row=1, col=2)
+
+        # third subplot
+        labels = ["boundary", "extent", "segment"]
+        values = [28.9, 37.6, 23.2]
+        trace_3 = go.Bar(
+            x=labels,
+            y=values,
+            orientation="v",
+            text=values,
+            width=[0.5, 0.5, 0.5],
+            textposition="outside",
+            marker_color=["moccasin", "lightgreen", "orangered"],
+        )
+        fig.add_trace(trace_3, row=1, col=3)
+
+        fig.update_layout(
+            height=400,
+            width=1200,
+            polar=dict(
+                radialaxis=dict(visible=True, showline=False, showticklabels=False, range=[0, 100])
+            ),
+            showlegend=False,
+            plot_bgcolor="rgba(0, 0, 0, 0)",
+            yaxis=dict(showticklabels=False),
+            yaxis_range=[0, int(max(values)) + 4],
+        )
+        fig.layout.annotations[0].update(y=1.2)
+        fig.layout.annotations[1].update(y=1.2)
+        fig.layout.annotations[2].update(y=1.2)
+
+        return fig
